all: add Order method to graph.Graph
This has significant potential to save useless work in collating node
when only the number of nodes is needed.

Order is a general property, not just of finite graphs, but all graphs
we have the capacity to deal with - our nodes have unique IDs, so there
is at most maxuint nodes even for a dynamically generated graph. Losing
half of these is not a significant issue.