{concrete,network,search,traverse}: allow graphs to specify absent edge cost
Note the rigmarole necessary to get around the fact that an edge cannot
carry its own edge weight. We can retain the current interface behaviour
of weighted graphs by having the graph access the edge weights for the
client, but still have the weights stored in the edge.

How to deal with SetEdgeWeight for absent edges? This should be a panic.