internal: Add Reset method to NodeStack
When doing breadth first traversal the NodeStack is reset by doing a direct
slice re-size on the underlying data structure (slice) used to implement
the stack. This violates the stack abstraction provided by NodeStack.
Instead of directly re-sizing the slice we can add a method to NodeStack to
reset the underlying data structure.

This patch adds Reset() method to NodeStack. This allows the stack to be
reset while maintaining data abstraction.